requirements.txt
PyQt5>=5.15.9
pyserial>=3.5
